Feat: LCFS - Refactor ViewTransfer and AddEditTransfer components #455
