+++ frontend/src/assets/locales/en/transfer.json
@@ -34,6 +34,7 @@
   "AgreementDt": "Agreement Date",
   "sendConfirmText": "Are you sure you want to sign and send this transfer to ",
   "submitConfirmText": "Are you sure you want to sign and submit this transfer to ",
+  "loadingText": "Loading transfer...",
   "creatingText": "Creating Draft Transfer...",
   "updatingText": "Updating Draft Transfer...",
   "processingText": "Processing Transfer...",
@@ -60,6 +61,9 @@
   "recordSuccessText": "Transfer successfully recorded.",
   "recordErrorText": "Error recording transfer",
   "refuseTransferBtn": "Refuse transfer",
+  "refuseConfirmText": "Are you sure you want to record this transfer?",
+  "refuseSuccessText": "Transfer successfully recorded.",
+  "refuseErrorText": "Error recording transfer",
   "returnToAnalystBtn": "Return to analyst",
   "returnWarningText": "You are strongly encouraged to provide an explanation to the analyst as to why the transfer is being returned.",
   "returnConfirmText": "Are you sure you want to return the transfer to the Analyst? This will change the status of the transfer back to Submitted.",
