+++ frontend/src/views/Transfers/components/Comments.jsx
@@ -5,10 +5,10 @@ import { useState } from 'react'
 import ExpandLessIcon from '@mui/icons-material/ExpandLess'
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore'
 import { useFormContext } from 'react-hook-form'
-import LabelBox from './LabelBox'
+import { LabelBox } from './LabelBox'
 import { useTranslation } from 'react-i18next'
 
-const Comments = () => {
+export const Comments = () => {
   const { t } = useTranslation(['transfer'])
   const [isExpanded, setIsExpanded] = useState(false)
 
@@ -49,5 +49,3 @@ const Comments = () => {
     </>
   )
 }
-
-export default Comments
