+++ frontend/src/views/Transfers/components/index.js
@@ -2,3 +2,11 @@ export { AttachmentList } from './AttachmentList';
 export { OrganizationBadge } from './OrganizationBadge';
 export { CommentList } from './CommentList';
 export { AddPlainComment } from './AddPlainComment';
+export { AgreementDate } from './AgreementDate';
+export { Comments } from './Comments';
+export { ConfirmationModal } from './ConfirmationModal';
+export { LabelBox } from './LabelBox';
+export { TransferDetails } from './TransferDetails';
+export { TransferDetailsCard } from './TransferDetailsCard';
+export { TransferGraphic } from './TransferGraphic';
+export { TransferSummary } from './TransferSummary';
