+++ frontend/src/views/Transfers/index.js
@@ -1,2 +1,1 @@
-export { AddEditTransfer } from './AddEditTransfer/AddEditTransfer'
-export { ViewTransfer } from './ViewTransfer/ViewTransfer'
+export { AddEditViewTransfer } from './AddEditViewTransfer'
